Rework ToolHelpMarkdown so fancy uris can be reused.
