+++ client/src/components/Markdown/gxuris-types.ts
@@ -1,0 +1,4 @@
+export interface HelpElementReference {
+    element: HTMLElement;
+    term: string;
+}
