+++ client/src/components/Tool/ToolHelpMarkdown.vue
@@ -1,11 +1,9 @@
 <script setup lang="ts">
-import { computed, onMounted, ref } from "vue";
+import { computed, ref } from "vue";
 
+import { useGxUris } from "@/components/Markdown/gxuris";
 import { markup } from "@/components/ObjectStore/configurationMarkdown";
 import { useFormattedToolHelp } from "@/composables/formattedToolHelp";
-import { getAppRoot } from "@/onload/loadConfig";
-
-import HelpPopover from "@/components/Help/HelpPopover.vue";
 
 const props = defineProps<{
     content: string;
@@ -18,38 +16,7 @@ const { formattedContent } = useFormattedToolHelp(markdownHtml);
 
 const helpHtml = ref<HTMLDivElement>();
 
-interface InternalTypeReference {
-    element: HTMLElement;
-    term: string;
-}
-
-const internalHelpReferences = ref<InternalTypeReference[]>([]);
-
-function setupPopovers() {
-    internalHelpReferences.value.length = 0;
-    if (helpHtml.value) {
-        const links = helpHtml.value.getElementsByTagName("a");
-        Array.from(links).forEach((link) => {
-            if (link.href.startsWith("gxhelp://")) {
-                const uri = link.href.substr("gxhelp://".length);
-                internalHelpReferences.value.push({ element: link, term: uri });
-                link.href = `${getAppRoot()}help/terms/${uri}`;
-                link.style.color = "inherit";
-                link.style.textDecorationLine = "underline";
-                link.style.textDecorationStyle = "dashed";
-            }
-        });
-        const imgs = helpHtml.value.getElementsByTagName("img");
-        Array.from(imgs).forEach((img) => {
-            if (img.src.startsWith("gxstatic://")) {
-                const rest = img.src.substr("gxstatic://".length);
-                img.src = `${getAppRoot()}static/${rest}`;
-            }
-        });
-    }
-}
-
-onMounted(setupPopovers);
+const { internalHelpReferences, MarkdownHelpPopovers } = useGxUris(helpHtml);
 </script>
 
 <template>
@@ -60,8 +27,6 @@ onMounted(setupPopovers);
         -->
         <!-- eslint-disable-next-line vue/no-v-html -->
         <div ref="helpHtml" v-html="formattedContent" />
-        <span v-for="(value, i) in internalHelpReferences" :key="i">
-            <HelpPopover :target="value.element" :term="value.term" />
-        </span>
+        <MarkdownHelpPopovers :elements="internalHelpReferences" />
     </span>
 </template>
